Attempt to speed startup by default on larger ring sizes
Make the default number of vnode to start per management tick higher.  Still start them in series, but prefer those not owned in the current ring - and so less likely to contain data.

This makes it more likely that the `maybe_ring_changed/4` function will start the active vnodes in parallel as expected.