Allow for negative filters
This allows for a riak_core_vnode_manager that is managing a repair to return negative filters when a second handoff is requested for the same target partition - so that data already sent in the first handoff is not resent.  If using a fold_heads, this will avoid the deferred fetch.