Improve logging at handoff completion
On handoff completion log the reason why a completion is being ignored (i.e. no ring update is being made).  Also log that a ring update has been triggered.

Also with the leveled backend quieten logs associated with monitoring of the leveled backend when unregistering.

The deletion of a backend may possibly be time-consuming when a system is under stress, and the vnode will be blocked for this time - so log the time it takes to delete the backend to help with monitoring.