Add support for {F, async} as a data_size option
This is intended for backends where it is convenient to return a function, (because the size is too expensive to calculate synchronously), but expensive to recalculate continuously (as would happen with {F, dynamic}, where the size is recalculated for every user CLI request for handoff/transfer status).

It would also be possible for the F() in {F, async} to return {F0, dynamic} should this be required.