Nhse o34 orc.i25 efficiency (#26)
* Simple, potential changes for efficiency

Always keep vclocks sorted - and use the sorting to speed-up comparison.

Also try to optimise the naming functions when we're naming riak_kv_vnode.

Still the outstanding issue of the long IndexName - commonly with about 150 0s - making the integer_to_list/list_to_binary conversion unnecessarily hard.

* Alternate - explicitly sort in vclock descends

* Pre-sort if dominates

As going to do two descends - so don't do two sorts

* Use persistent term rather than recalculate name

We're going to look this up over and over - so don't recalculate

* Make non-legacy routing permanent

* Confirm performance in speed test

Keep test to use for future potential refactoring

* Minor edits

Speed up for options fetching, but just on metadata get.

The vclock will now support a future when the bucket properties is a map - i.e because a map is passed around in riak_kv instead of a list.